[SPARK-25332][SQL] Instead of broadcast hash join ,Sort merge join has selected when restart spark-shell/spark-JDBC for hive provider
## What changes were proposed in this pull request?
Problaem:
Below steps in sequence to reproduce the issue.
a.Create parquet table with stored as clause.
b.Run insert statement  => This will not update Hivestats.
c.Run (Select query which needs to calculate stats or explain cost select statement)  -> this will evaluate  stats from HadoopFsRelation
d.Since correct stats(sizeInBytes) is available , the plan  will select  broadcast node if join with any table.
e. Exit; (come out of shell)

f.Now again run setp c ( calculate stat) query. This gives wrong stats (sizeInBytes - default value will be assigned)  in plan. Because it is calculated from HiveCatalogTable which does not have any stats as it is not updated in step b
g.Since in-correct stats(sizeInBytes - default value will be assigned) is available, the plan  will select  SortMergeJoin node if join with any table.
h.Now Run insert statement => This will  update Hivestats .
i.Now again run setp c ( calculate stat) query. This gives correct stat (sizeInBytes)  in plan .because it can read the hive stats which is updated in step i.
j.Now onward always stat is available so correct stat is plan will be displayed  which picks  Broadcast join node(based on threshold size) always.

Solution:
Here the main problem is hive stats is not getting recorded after insert command, this is because of a condition  "if (table.stats.nonEmpty)" in updateTableStats()  which will be executed as part of InsertIntoHadoopFsRelationCommand command.
So as part of fix we initialized a default value for the CatalogTable stats if there is no cache of a particular LogicalRelation.

Also it is observed in Test Case “test statistics of LogicalRelation converted from Hive serde tables" in StatisticsSuite,  orc and parquet both are convertible but we are expecting that only orc should
get/record stats Hivestats  not for parquet.But since both relations are convertible so we should have same expectation. Same is corrected in this PR.

How this patch tested:
Manually testes, adding the test snapshots and the UT is corrected which will verify the PR scenario.
